create-list-role

+++ routes/web.php
@@ -15,18 +15,15 @@ Route::get('/', function () {
     return view('welcome');
 });
 
-Route::get('/admin', function() {
-  return view('admin.demo');
-});
-
 Route::group(['prefix'=>'admin'], function(){
+  Route::get('/', 'AdminController@index')->name('admin.index');
   Route::group(['prefix'=>'roles'], function(){
-    Route::get('/', 'RoleController@index');
+    Route::get('/', 'RoleController@index')->name('role.index');
     Route::get('/create', 'RoleController@create');
     Route::post('/', 'RoleController@store');
     Route::get('/{id}', 'RoleController@show');
     Route::get('/{id}/edit', 'RoleController@edit');
     Route::put('/{id}', 'RoleController@update');
-    Route::delete('/{id}', 'RoleController@destroy');
+    Route::delete('/{id}', 'RoleController@destroy')->name('role.destroy');
   });
 });
